# openapi/gateway.openapi.json
{
  "components": {
    "schemas": {
      "AgentRequest": {
        "description": "Input to ``POST /v1/agent`` (Wave 7).",
        "example": {
          "max_steps": 4,
          "question": "summarise the README",
          "session_id": "user-42"
        },
        "properties": {
          "max_steps": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-request override of the AGENT_MAX_STEPS budget.",
            "title": "Max Steps"
          },
          "question": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "If set, the agent loads the session's prior turns and persists the new pair.",
            "title": "Session Id"
          },
          "use_chat_history": {
            "default": true,
            "description": "If false, session_id is ignored (stateless turn).",
            "title": "Use Chat History",
            "type": "boolean"
          }
        },
        "required": [
          "question"
        ],
        "title": "AgentRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QueryRequest": {
        "example": {
          "question": "What does foo return?",
          "session_id": "user-42",
          "top_k": 5
        },
        "properties": {
          "question": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "If set, prior turns in this session are sent to the model and updated after each reply.",
            "title": "Session Id"
          },
          "top_k": {
            "default": 5,
            "description": "Number of KB chunks to merge into context",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          },
          "use_chat_history": {
            "default": true,
            "description": "If false, session_id is ignored (stateless turn).",
            "title": "Use Chat History",
            "type": "boolean"
          }
        },
        "required": [
          "question"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "BFF orchestrating ingest and RAG microservices",
    "title": "sorakAi Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agent/stream": {
      "post": {
        "operationId": "proxy_agent_stream_v1_agent_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proxy Agent Stream",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/v1/query/stream": {
      "post": {
        "operationId": "proxy_query_stream_v1_query_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proxy Query Stream",
        "tags": [
          "rag"
        ]
      }
    }
  },
...
}